requirements.txt
langchain>=0.0.300
openai>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
feature-engine>=1.6.0
sweetviz>=2.2.0
ydata-profiling>=4.6.0
mlflow>=2.8.0
prefect>=2.13.0
gradio>=4.0.0
jupyter>=1.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
joblib>=1.3.0
plotly>=5.18.0
shap>=0.43.0
optuna>=3.4.0
great-expectations>=0.17.19
pytest>=7.4.0
black>=23.10.0
isort>=5.12.0
flake8>=6.1.0 